Add description template to accomplishment types
I was recently shown a good template for keeping track of accomplishments at work -- a.k.a. "Hype Sheet" format. I didn't need to build in a bunch of hype sheet architecture to accomplish this. I chose to implement markdown rendering (#18) and a template for specific types of entries. Now when I choose a "Merit" type entry, the description will populate with the "hype sheet" format.